KVM: TDX: Allow TD to create L2 TD partitioning guests
The userspace VMM can determine the number of L2 VMs supported by a
single TD through the exposed TDX capabilities. A new flag has been
introduced in kvm_tdx_init_vm, allowing the userspace VMM to specify the
exact number of L2 VMs it requires. This value is then set to the new
field NUM_L2_VMS in td_params for TDX module to reference.

With NUM_L2_VMS set in td_params, each L2 VM requires additional pages
for TDCS/TDVPS. The number of tdcs_pages and tdvpx_pages are stored in
the kvm_tdx on a per-VM basis, and pages are allocated accordingly.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>